oop solutions done, still need tester
